requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --python 3.11 requirements.in -o requirements.txt
accelerate==1.5.2
    # via peft
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.16
    # via torch-geometric
aiosignal==1.3.2
    # via aiohttp
alembic==1.15.2
    # via optuna
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via omegaconf
anyio==4.8.0
    # via httpx
argon2-cffi==23.1.0
    # via minio
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
attrs==25.1.0
    # via aiohttp
blessed==1.20.0
    # via gpustat
carte-ai==0.0.26
    # via -r requirements.in
catboost==1.2.7
    # via
    #   -r requirements.in
    #   carte-ai
category-encoders==2.7.0
    # via carte-ai
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   minio
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via wandb
colorlog==6.9.0
    # via optuna
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
docker-pycreds==0.4.0
    # via wandb
einops==0.8.1
    # via tabpfn
fasttext==0.9.3
    # via carte-ai
filelock==3.17.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.56.0
    # via matplotlib
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.2.0
    # via
    #   huggingface-hub
    #   torch
    #   torch-geometric
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via wandb
gpustat==1.1.1
    # via -r requirements.in
graphviz==0.20.3
    # via catboost
h11==0.14.0
    # via httpcore
h5py==3.13.0
    # via -r requirements.in
httpcore==1.0.7
    # via httpx
httpx==0.27.2
    # via tabpfn-client
huggingface-hub==0.29.1
    # via
    #   -r requirements.in
    #   accelerate
    #   peft
    #   sentence-transformers
    #   tabpfn
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via
    #   skrub
    #   torch
    #   torch-geometric
joblib==1.4.2
    # via scikit-learn
kagglehub==0.3.11
    # via -r requirements.in
kiwisolver==1.4.8
    # via matplotlib
liac-arff==2.5.0
    # via openml
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
matplotlib==3.10.1
    # via
    #   catboost
    #   seaborn
    #   skrub
minio==7.2.15
    # via openml
mpmath==1.3.0
    # via sympy
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
narwhals==1.28.0
    # via plotly
networkx==3.4.2
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   accelerate
    #   carte-ai
    #   catboost
    #   category-encoders
    #   contourpy
    #   fasttext
    #   h5py
    #   matplotlib
    #   openml
    #   optuna
    #   pandas
    #   patsy
    #   peft
    #   scikit-learn
    #   scipy
    #   seaborn
    #   skrub
    #   statsmodels
    #   tabpfn-extensions
    #   torch-geometric
    #   torchvision
    #   transformers
    #   xgboost
nvidia-ml-py==12.570.86
    # via gpustat
omegaconf==2.3.0
    # via tabpfn-client
openml==0.15.1
    # via -r requirements.in
optuna==4.3.0
    # via -r requirements.in
packaging==24.2
    # via
    #   accelerate
    #   huggingface-hub
    #   kagglehub
    #   matplotlib
    #   openml
    #   optuna
    #   peft
    #   plotly
    #   pytest
    #   skrub
    #   statsmodels
    #   transformers
pandas==2.2.3
    # via
    #   -r requirements.in
    #   carte-ai
    #   catboost
    #   category-encoders
    #   openml
    #   seaborn
    #   skrub
    #   statsmodels
    #   tabpfn
    #   tabpfn-client
    #   tabpfn-extensions
password-strength==0.0.3.post2
    # via tabpfn-client
patsy==1.0.1
    # via
    #   category-encoders
    #   statsmodels
peft==0.15.1
    # via -r requirements.in
pillow==11.1.0
    # via
    #   matplotlib
    #   sentence-transformers
    #   torchvision
platformdirs==4.3.6
    # via wandb
plotly==6.0.0
    # via catboost
pluggy==1.5.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.3
    # via wandb
psutil==7.0.0
    # via
    #   accelerate
    #   gpustat
    #   peft
    #   torch-geometric
    #   wandb
pyarrow==19.0.1
    # via
    #   carte-ai
    #   openml
pybind11==2.13.6
    # via fasttext
pycparser==2.22
    # via cffi
pycryptodome==3.21.0
    # via minio
pydantic==2.10.6
    # via wandb
pydantic-core==2.27.2
    # via pydantic
pyparsing==3.2.1
    # via
    #   matplotlib
    #   torch-geometric
pytest==8.3.4
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   openml
    #   pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   kagglehub
    #   omegaconf
    #   optuna
    #   peft
    #   transformers
    #   wandb
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   kagglehub
    #   openml
    #   skrub
    #   torch-geometric
    #   transformers
    #   wandb
safetensors==0.5.3
    # via
    #   accelerate
    #   peft
    #   transformers
scikit-learn==1.5.2
    # via
    #   -r requirements.in
    #   carte-ai
    #   category-encoders
    #   openml
    #   sentence-transformers
    #   skrub
    #   tabpfn
    #   tabpfn-client
    #   tabpfn-extensions
scipy==1.15.2
    # via
    #   -r requirements.in
    #   carte-ai
    #   catboost
    #   category-encoders
    #   openml
    #   scikit-learn
    #   sentence-transformers
    #   skrub
    #   statsmodels
    #   tabpfn
    #   tabpfn-extensions
    #   xgboost
seaborn==0.12.2
    # via
    #   carte-ai
    #   tabpfn-extensions
sentence-transformers==3.4.1
    # via -r requirements.in
sentry-sdk==2.22.0
    # via wandb
setproctitle==1.3.5
    # via wandb
setuptools==75.8.2
    # via
    #   fasttext
    #   wandb
six==1.17.0
    # via
    #   blessed
    #   catboost
    #   docker-pycreds
    #   password-strength
    #   python-dateutil
skrub==0.5.1
    # via -r requirements.in
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.40
    # via
    #   alembic
    #   optuna
sseclient-py==1.8.0
    # via tabpfn-client
statsmodels==0.14.4
    # via category-encoders
sympy==1.13.1
    # via torch
tabpfn==2.0.8
    # via
    #   -r requirements.in
    #   carte-ai
tabpfn-client==0.1.7
    # via -r requirements.in
tabpfn-extensions==0.0.4
    # via -r requirements.in
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
torch==2.6.0
    # via
    #   -r requirements.in
    #   accelerate
    #   carte-ai
    #   peft
    #   sentence-transformers
    #   tabpfn
    #   tabpfn-extensions
    #   torchvision
torch-geometric==2.6.1
    # via carte-ai
torcheval==0.0.7
    # via carte-ai
torchvision==0.21.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   kagglehub
    #   openml
    #   optuna
    #   peft
    #   sentence-transformers
    #   tabpfn-client
    #   torch-geometric
    #   transformers
transformers==4.49.0
    # via
    #   -r requirements.in
    #   peft
    #   sentence-transformers
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   huggingface-hub
    #   minio
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   tabpfn
    #   torch
    #   torcheval
    #   wandb
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via
    #   minio
    #   requests
    #   sentry-sdk
wandb==0.19.7
    # via -r requirements.in
wcwidth==0.2.13
    # via blessed
xgboost==2.1.4
    # via
    #   -r requirements.in
    #   carte-ai
xmltodict==0.14.2
    # via openml
xxhash==3.5.0
    # via tabpfn-client
yarl==1.19.0
    # via aiohttp
